+++ pynonymizer/pynonymize.py
@@ -10,9 +10,13 @@ from pynonymizer.process_steps import StepActionMap, ProcessSteps
 logger = get_default_logger()
 
 
-def pynonymize(input_path=None, strategyfile_path=None, output_path=None, db_user=None, db_password=None, db_type=None,
-               db_host=None, db_name=None, fake_locale=None, start_at_step=None, stop_at_step=None, skip_steps=None,
-               seed_rows=None):
+def pynonymize(
+        input_path=None, strategyfile_path=None, output_path=None, db_user=None, db_password=None, db_type=None,
+        db_host=None, db_name=None, fake_locale=None, start_at_step=None, stop_at_step=None, skip_steps=None,
+        seed_rows=None,
+
+        **kwargs
+    ):
 
     # Default and Normalize args
     if start_at_step is None:
@@ -72,6 +76,15 @@ def pynonymize(input_path=None, strategyfile_path=None, output_path=None, db_use
 
 
     # Initialize and validate DB credentials (always required)
+
+    # Discover db-type kwargs
+    # mssql_backup_option -> backup_option and pass these to the constructor
+    db_kwargs = {}
+    db_arg_prefix = f"{db_type}_"
+    for k, v in kwargs.items():
+        if k.startswith(db_arg_prefix):
+            db_kwargs[ k[len(db_arg_prefix):] ] = v
+
     logger.debug("Database: (%s)%s@%s db_name: %s", db_host, db_type, db_user, db_name)
     db_provider = get_provider(
         type=db_type,
@@ -79,7 +92,8 @@ def pynonymize(input_path=None, strategyfile_path=None, output_path=None, db_use
         db_user=db_user,
         db_pass=db_password,
         db_name=db_name,
-        seed_rows=seed_rows
+        seed_rows=seed_rows,
+        **db_kwargs
     )
 
     if not db_provider.test_connection():